item

+++ Server/src/main/java/com/auction/server/repository/ItemRepository.java
@@ -2,8 +2,9 @@ package com.auction.server.repository;
 
 import com.auction.server.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item,Long> {
+public interface ItemRepository extends JpaRepository<Item,Integer> {
 }
